Backend/requirements.txt
flask
flask-cors
pandas
python-dotenv
google-generativeai
speechrecognition
pydub
twilio